change int64 to usize, remove Option<span>

+++ rust/src/control_flow/call_binding.rs
@@ -2,6 +2,6 @@ use crate::types::types::Type;
 
 #[derive(Debug, Clone)]
 pub struct CallBinding {
-    pub symbol_id: i64,
+    pub symbol_id: usize,
     pub ty: Type,
 }
