+++ rust/src/control_flow/class_cfg.rs
@@ -2,11 +2,11 @@ use crate::control_flow::graph::Graph;
 
 pub struct ClassCfg<'a> {
     pub graph: Graph<'a>,
-    pub scope_id: i64,
+    pub scope_id: usize,
 }
 
 impl<'a> ClassCfg<'a> {
-    pub fn new(scope_id: i64) -> Self {
+    pub fn new(scope_id: usize) -> Self {
         Self {
             graph: Graph::new(),
             scope_id,
