+++ rust/src/control_flow/function_cfg.rs
@@ -7,11 +7,11 @@ pub struct FunctionCfg<'a> {
     pub graph: Graph<'a>,
     pub params: Vec<ArgIR>,
     pub returns: Option<ExprIR>,
-    pub scope_id: i64,
+    pub scope_id: usize,
 }
 
 impl<'a> FunctionCfg<'a> {
-    pub fn new(params: Vec<ArgIR>, returns: Option<ExprIR>, scope_id: i64) -> Self {
+    pub fn new(params: Vec<ArgIR>, returns: Option<ExprIR>, scope_id: usize) -> Self {
         Self {
             graph: Graph::new(),
             params,
