+++ rust/src/control_flow/module_cfg.rs
@@ -2,5 +2,5 @@ use crate::control_flow::graph::Graph;
 
 pub struct ModuleCfg<'a> {
     pub graph: Graph<'a>,
-    pub scope_id: i64,
+    pub scope_id: usize,
 }
